requirements.txt
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@afe9eb920646102f7e6bf0cd2115841cea2aca13
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.14.3
opencv-python==4.9.0.80
pandas==2.1.4
scipy==1.11.4
segment-anything==1.0
streamlit==1.27.2
streamlit-drawable-canvas==0.9.3
torch==2.2.1
torchmetrics==1.3.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.39.2
anthropic
python-dotenv